Fix a double conversion util to not lost precision.
Apparently copy&paste bug where a float temp variable was used.

So it means all p-values until today were processed as floats 8-)

+++ rtt/utils.cpp
@@ -49,7 +49,7 @@ double Utils::strtod(const std::string & str) {
                                  "string contain invalid characters");
     }
     try {
-        float result = std::stod(str);
+        double result = std::stod(str);
         return result;
     } catch (std::out_of_range) {
         /* invalid_argument won't be thrown, regex prevents that */
